Moved all Hyper-V management code into `HyperVHelper`.
