+++ src/Hyper-V Tray/VirtualMachineStateChangedEventArgs.cs
@@ -1,0 +1,26 @@
+﻿using System;
+
+namespace HyperVTray
+{
+    internal class VirtualMachineStateChangedEventArgs : EventArgs
+    {
+        #region Construction
+
+        public VirtualMachineStateChangedEventArgs(string name, VmState state, bool critical)
+        {
+            Critical = critical;
+            Name = name;
+            State = state;
+        }
+
+        #endregion
+
+        #region Properties
+
+        public bool Critical { get; }
+        public string Name { get; }
+        public VmState State { get; }
+
+        #endregion
+    }
+}
